requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Limiter==1.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
